Install Firefox on desktop to make Geckodriver work.
Hmm in the past there's been a bug when Geckodriver can't talk
to Android if not Firefox is installed. Looking at the errors we
get and looking in Selenium source it looks like it looking for
Firefox on the server. I remember this upstream loong time ago but
I didn't file a task. But still, something must have changed at
BitBar.

Bug: T331994
Change-Id: Ibd17eb85bd1cbe0dd6c862174c6c12a227e845ac