New structure for BitBar tests.
This will introduce a new way to run the test. The server that kicks
off the tests, will run bitbarcaller.sh $PARAM that will start to send
jobs
to bitbar by looking in the $PARAM test folder. The jobs will be queued
at bitbar and handled one by one.

To add a new test, you add it to the folder.

For a start we will kick off the jobs every 3 hours in the crontab. There
are some work that need to be done so that Chrome/Firefox jobs are split
but that is something I can do in a follow up.

Bug: T330402
Change-Id: Ib73999bd5426412145a9e228c659b15e1edff015